Terminado añadir y borrar partidos

+++ ffb_daw_fase3/src/main/java/com/ffbet/fase3/domain/Match.java
@@ -45,9 +45,9 @@ public abstract class Match {
 	@Column(nullable = false)
 	protected String visitingTeam;
 	@Column(nullable = false)
-	protected int quotaHomeVictory;
+	protected double quotaHomeVictory;
 	@Column(nullable = false)
-	protected int quotaVisitingVictory;
+	protected double quotaVisitingVictory;
 	@ManyToMany
 	protected List<Team> teams = new ArrayList<>();
 	
@@ -78,8 +78,8 @@ public abstract class Match {
 	 * @param (required)
 	 *            quotaVisitingVictory, the quota of visiting victory's bet
 	 */
-	public Match(Date date, Time time, String homeTeam, String visitingTeam, int quotaHomeVictory,
-			int quotaVisitingVictory) {
+	public Match(Date date, Time time, String homeTeam, String visitingTeam, double quotaHomeVictory,
+			double quotaVisitingVictory) {
 		super();
 		this.date = date;
 		this.time = time;
@@ -131,19 +131,19 @@ public abstract class Match {
 		this.visitingTeam = visitingTeam;
 	}
 
-	public int getQuotaHomeVictory() {
+	public double getQuotaHomeVictory() {
 		return quotaHomeVictory;
 	}
 
-	public void setQuotaHomeVictory(int quotaHomeVictory) {
+	public void setQuotaHomeVictory(double quotaHomeVictory) {
 		this.quotaHomeVictory = quotaHomeVictory;
 	}
 
-	public int getQuotaVisitingVictory() {
+	public double getQuotaVisitingVictory() {
 		return quotaVisitingVictory;
 	}
 
-	public void setQuotaVisitingVictory(int quotaVisitingVictory) {
+	public void setQuotaVisitingVictory(double quotaVisitingVictory) {
 		this.quotaVisitingVictory = quotaVisitingVictory;
 	}
